+++ Assets/Plugins/Sirenix/Odin Inspector/Config/Editor/ColorPaletteManager.asset
@@ -1,0 +1,136 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 772478971, guid: a4865f1ab4504ed8a368670db22f409c, type: 3}
+  m_Name: ColorPaletteManager
+  m_EditorClassIdentifier: 
+  colorPalettes:
+  - name: Country
+    showAlpha: 0
+    colors:
+    - {r: 0.776, g: 0.651, b: 0.349, a: 1}
+    - {r: 0.863, g: 0.761, b: 0.631, a: 1}
+    - {r: 0.91, g: 0.831, b: 0.686, a: 1}
+    - {r: 0.961, g: 0.902, b: 0.788, a: 1}
+    - {r: 0.753, g: 0.714, b: 0.667, a: 1}
+    - {r: 0.478, g: 0.573, b: 0.431, a: 1}
+    - {r: 0.314, g: 0.427, b: 0.31, a: 1}
+    - {r: 0.596, g: 0.345, b: 0.235, a: 1}
+    - {r: 0.545, g: 0.329, b: 0.318, a: 1}
+    - {r: 0.647, g: 0.204, b: 0.227, a: 1}
+    - {r: 0.435, g: 0.161, b: 0.063, a: 1}
+    - {r: 0.357, g: 0.333, b: 0.278, a: 1}
+    - {r: 0.976, g: 0.98, b: 0.961, a: 1}
+    - {r: 0.165, g: 0.271, b: 0.11, a: 1}
+  - name: Beach
+    showAlpha: 0
+    colors:
+    - {r: 0.996, g: 0.906, b: 0.459, a: 1}
+    - {r: 0.314, g: 0.592, b: 0.035, a: 1}
+    - {r: 0.486, g: 0.953, b: 0.875, a: 1}
+    - {r: 0.996, g: 0.82, b: 0.212, a: 1}
+    - {r: 1, g: 0.769, b: 0.165, a: 1}
+    - {r: 0.804, g: 0.835, b: 0.753, a: 1}
+    - {r: 1, g: 0.769, b: 0.165, a: 1}
+    - {r: 1, g: 0.702, b: 0.063, a: 1}
+    - {r: 1, g: 0.898, b: 0.569, a: 1}
+  - name: Fall
+    showAlpha: 0
+    colors:
+    - {r: 0.82, g: 0.722, b: 0.318, a: 1}
+    - {r: 0.537, g: 0.192, b: 0.153, a: 1}
+    - {r: 0.996, g: 0.812, b: 0.012, a: 1}
+    - {r: 1, g: 0.431, b: 0.02, a: 1}
+    - {r: 0.937, g: 0.267, b: 0.094, a: 1}
+    - {r: 0.42, g: 0.212, b: 0.18, a: 1}
+    - {r: 0.992, g: 0.651, b: 0.004, a: 1}
+    - {r: 0.89, g: 0.353, b: 0.086, a: 1}
+    - {r: 1, g: 0.443, b: 0.004, a: 1}
+    - {r: 0.682, g: 0.275, b: 0.137, a: 1}
+    - {r: 0.306, g: 0.231, b: 0.114, a: 1}
+    - {r: 0.384, g: 0.416, b: 0.082, a: 1}
+    - {r: 0.165, g: 0.157, b: 0.008, a: 1}
+    - {r: 0.906, g: 0.635, b: 0.227, a: 1}
+    - {r: 0.82, g: 0.722, b: 0.318, a: 1}
+    - {r: 0.745, g: 0.435, b: 0.031, a: 1}
+    - {r: 0.765, g: 0.682, b: 0.569, a: 1}
+    - {r: 0.18, g: 0.149, b: 0.075, a: 1}
+    - {r: 0.702, g: 0.451, b: 0.059, a: 1}
+  - name: Passion
+    showAlpha: 0
+    colors:
+    - {r: 0.925, g: 0.682, b: 0.624, a: 1}
+    - {r: 0.188, g: 0.114, b: 0.224, a: 1}
+    - {r: 0.349, g: 0.11, b: 0.231, a: 1}
+    - {r: 0.435, g: 0.267, b: 0.357, a: 1}
+  - name: Sepia
+    showAlpha: 0
+    colors:
+    - {r: 0.353, g: 0.098, b: 0.02, a: 1}
+    - {r: 0.663, g: 0.188, b: 0.114, a: 1}
+    - {r: 0.906, g: 0.643, b: 0.082, a: 1}
+    - {r: 0.996, g: 0.839, b: 0.322, a: 1}
+    - {r: 0.486, g: 0.392, b: 0.02, a: 1}
+    - {r: 0.294, g: 0.235, b: 0.012, a: 1}
+  - name: Floral
+    showAlpha: 0
+    colors:
+    - {r: 0.855, g: 0.518, b: 0.412, a: 1}
+    - {r: 0.827, g: 0.294, b: 0.333, a: 1}
+    - {r: 0.737, g: 0.118, b: 0.208, a: 1}
+    - {r: 0.549, g: 0.149, b: 0.235, a: 1}
+    - {r: 0.949, g: 0.925, b: 0.784, a: 1}
+    - {r: 0.945, g: 0.882, b: 0.69, a: 1}
+    - {r: 0.871, g: 0.812, b: 0.698, a: 1}
+    - {r: 0.4, g: 0.196, b: 0.243, a: 1}
+    - {r: 0.271, g: 0.157, b: 0.227, a: 1}
+  - name: Underwater
+    showAlpha: 0
+    colors:
+    - {r: 0.663, g: 0.416, b: 0.733, a: 1}
+    - {r: 0.2, g: 0.6, b: 0.698, a: 1}
+    - {r: 0.11, g: 0.49, b: 0.698, a: 1}
+    - {r: 0.439, g: 0.627, b: 0.227, a: 1}
+    - {r: 0, g: 0.357, b: 0.604, a: 1}
+    - {r: 0.067, g: 0.271, b: 0.353, a: 1}
+  - name: Breeze
+    showAlpha: 0
+    colors:
+    - {r: 0.706, g: 1, b: 0, a: 1}
+    - {r: 0.651, g: 1, b: 0.404, a: 1}
+    - {r: 0.122, g: 1, b: 0.514, a: 1}
+    - {r: 0.216, g: 0.894, b: 0.961, a: 1}
+    - {r: 0.4, g: 1, b: 0.882, a: 1}
+    - {r: 0.027, g: 0.792, b: 0.8, a: 1}
+  - name: Clovers
+    showAlpha: 0
+    colors:
+    - {r: 0.431, g: 0.549, b: 0.102, a: 1}
+    - {r: 0.671, g: 0.714, b: 0.071, a: 1}
+    - {r: 0.969, g: 0.949, b: 0.831, a: 1}
+    - {r: 0.886, g: 0.902, b: 0.702, a: 1}
+    - {r: 0.753, g: 0.824, b: 0.627, a: 1}
+    - {r: 0.404, g: 0.6, b: 0.4, a: 1}
+  - name: Tropical
+    showAlpha: 0
+    colors:
+    - {r: 0.953, g: 0.647, b: 0.804, a: 1}
+    - {r: 0.965, g: 0.741, b: 0.871, a: 1}
+    - {r: 0.949, g: 0.549, b: 0.643, a: 1}
+    - {r: 0.992, g: 0.659, b: 0.498, a: 1}
+    - {r: 0.976, g: 0.792, b: 0.729, a: 1}
+    - {r: 0.984, g: 0.855, b: 0.725, a: 1}
+    - {r: 0.259, g: 0.882, b: 0.663, a: 1}
+    - {r: 0.349, g: 0.753, b: 0.78, a: 1}
+    - {r: 0.725, g: 0.976, b: 0.91, a: 1}
+    - {r: 0.647, g: 0.745, b: 0.957, a: 1}
+    - {r: 0.725, g: 0.863, b: 0.973, a: 1}
+    - {r: 0.89, g: 0.945, b: 0.996, a: 1}
