+++ Assets/Plugins/Sirenix/Odin Inspector/Config/Editor/ColorPaletteManager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1fc564c78a50fbb4196e0eb56071a8a9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
