+++ Assets/Resources/Fonts/HeirofLightRegular SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b7e3bf7f2e248b045b097da13d765d2c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
