+++ Assets/Resources/Fonts/OK CHAN SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0654dfd52b7e81443875ef25b5f3c89d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
