+++ Assets/Resources/SO/Chicken.asset
@@ -14,10 +14,9 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   race: "\uB2ED"
   hpMax: 5
-  speedIdle: 0
-  speedRun: 0
+  speedIdle: 2
   speedAttack: 0
   damage: 0
-  attackSpeed: 0
+  attackCooldown: 0
   recognitionRange: 0
   attackRange: 0
