+++ Assets/Resources/SO/Rabbit.asset
@@ -1,0 +1,22 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5c2484e5d8c2a8149b1ad67e0e1075f0, type: 3}
+  m_Name: Rabbit
+  m_EditorClassIdentifier: 
+  race: "\uD1A0\uB07C"
+  hpMax: 7
+  speedIdle: 3
+  speedAttack: 0
+  damage: 0
+  attackCooldown: 0
+  recognitionRange: 0
+  attackRange: 0
