+++ Assets/Resources/SO/Rabbit.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26e693df579625a41bc0c65e0a20c18e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
