--- Assets/Scripts/DroppedItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 827da7d438f7ac348a2b18deb645b4bb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
