--- Assets/Scripts/Manager/ItemSpawner.cs
@@ -1,18 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ItemSpawner : Singleton<ItemSpawner>
-{
-	#region PublicVariables
-	#endregion
-
-	#region PrivateVariables
-	#endregion
-
-	#region PublicMethod
-	#endregion
-
-	#region PrivateMethod
-	#endregion
-}
