--- Assets/Scripts/Manager/ItemSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d886a0b240db3ed43a6cae5275a2d9c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
